CookiePersister, CookieStoreM7R

+++ src/main/java/org/mentalizr/cli/util/stringFilePersister/StringFilePersister.java
@@ -31,12 +31,31 @@ public class StringFilePersister {
         Files.writeString(this.path, string, this.charset, StandardOpenOption.CREATE, StandardOpenOption.TRUNCATE_EXISTING);
     }
 
+    /**
+     * Returns content from file as string. Existence of file is a precondition.
+     * If checking for file existence and reading can not be performed as an thread save (atomic) operation,
+     * catch IllegalStateException.
+     *
+     * @return content from file as String.
+     * @throws IOException
+     */
     public synchronized String read() throws IOException {
+        if (!exists()) throw new IllegalStateException("No such file to read from: " + this.path.toString());
         return Files.readString(this.path, this.charset);
     }
 
+    /**
+     * Deletes file. Existence of file is a precondition.
+     *
+     * @throws IOException
+     */
     public synchronized void delete() throws IOException {
+        if (!exists()) throw new IllegalStateException("No such file to delete: " + this.path.toString());
         Files.delete(this.path);
+    }
+
+    public synchronized void deleteIfExists() throws IOException {
+        Files.deleteIfExists(this.path);
     }
 
     public synchronized boolean exists() {
